# services/registration-analytics-service/api-spec.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Registration & Analytics Service API",
    "description": "Comprehensive service for managing event registrations, payment processing, \nattendance tracking, and real-time analytics with A/B testing capabilities.\n\n\nFeatures:\n- Complex registration workflows with custom form handling\n- Multi-tier ticketing system with dynamic pricing\n- Real-time analytics and time-series data\n- Payment processing integration\n- Attendance tracking with multiple check-in methods\n- A/B testing framework for event optimization\n- Advanced reporting and metrics aggregation\n",
    "version": "1.0.0",
    "contact": {
      "name": "Event Platform Team",
      "email": "api-support@eventplatform.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.eventplatform.com/registration/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.eventplatform.com/registration/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8083/api/v1",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/registrations/{registrationId}/checkins": {
      "post": {
        "summary": "Record a check-in",
        "description": "Record attendee check-in using various methods (QR code, NFC, manual, etc.).\nSupports multiple check-ins for multi-day events.\n",
        "tags": [
          "Check-ins"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RegistrationIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Check-in recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInRecord"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "Check-in conflict (already checked in, invalid status)"
          }
        }
      },
      "get": {
        "summary": "Get check-in history",
        "tags": [
          "Check-ins"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RegistrationIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Check-in history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheckInRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/events": {
      "post": {
        "summary": "Track analytics event",
        "description": "Record an analytics event for tracking user behavior and conversion metrics.\nSupports custom properties and automatic A/B test attribution.\n",
        "tags": [
          "Analytics"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsEventRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Analytics event accepted for processing"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/analytics/realtime/{eventId}": {
      "get": {
        "summary": "Get real-time analytics",
        "description": "Get real-time analytics data for an ongoing event including\ncurrent check-ins, session attendance, and live metrics.\n",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Real-time analytics data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealTimeAnalytics"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "parameters": {
...
      "RegistrationIdParam": {
        "name": "registrationId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "schemas": {
...
      "CheckInRecord": {
        "type": "object",
        "required": [
          "id",
          "checkInTime",
          "method",
          "location"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "checkInTime": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "$ref": "#/components/schemas/CheckInMethod"
          },
          "location": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "additionalData": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "CheckInMethod": {
        "type": "string",
        "enum": [
          "QR_CODE",
          "NFC",
          "MANUAL",
          "FACIAL_RECOGNITION",
          "RFID",
          "MOBILE_APP"
        ]
      },
      "CheckInRequest": {
        "type": "object",
        "required": [
          "method",
          "location"
        ],
        "properties": {
          "method": {
            "$ref": "#/components/schemas/CheckInMethod"
          },
          "location": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "verificationData": {
            "type": "object",
            "additionalProperties": true,
            "description": "QR code data, NFC tag info, etc."
          },
          "additionalData": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "AnalyticsEventRequest": {
        "type": "object",
        "required": [
          "eventType",
          "timestamp"
        ],
        "properties": {
          "eventId": {
            "type": "string"
          },
          "eventType": {
            "$ref": "#/components/schemas/AnalyticsEventType"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": true
          },
          "userAgent": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string",
            "format": "ipv4"
          },
          "location": {
            "$ref": "#/components/schemas/GeoPoint"
          }
        }
      },
      "AnalyticsEventType": {
        "type": "string",
        "enum": [
          "PAGE_VIEW",
          "REGISTRATION_START",
          "REGISTRATION_COMPLETE",
          "CHECK_IN",
          "SESSION_ATTENDANCE",
          "FEEDBACK_SUBMITTED",
          "PAYMENT_INITIATED",
          "PAYMENT_COMPLETED",
          "EMAIL_OPENED",
          "EMAIL_CLICKED"
        ]
      },
...
      "RealTimeAnalytics": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "currentAttendees": {
            "type": "integer"
          },
          "totalCheckIns": {
            "type": "integer"
          },
          "liveMetrics": {
            "type": "object",
            "properties": {
              "registrationsLastHour": {
                "type": "integer"
              },
              "checkInsLastHour": {
                "type": "integer"
              },
              "averageSessionDuration": {
                "type": "number",
                "format": "double"
              }
            }
          },
          "sessionAttendance": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "sessionId": {
                  "type": "string"
                },
                "attendeeCount": {
                  "type": "integer"
                },
                "capacity": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
...
      "GeoPoint": {
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "minimum": -90,
            "maximum": 90,
            "example": 40.7128
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "minimum": -180,
            "maximum": 180,
            "example": -74.006
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message",
          "timestamp",
          "path"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters or body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
...
}